Use int16_t for consistency with the params

+++ include/clap/ext/note-name.h
@@ -1,6 +1,5 @@
 #pragma once
 
-
 #include "../plugin.h"
 #include "../string-sizes.h"
 
@@ -13,10 +12,10 @@ extern "C" {
 static CLAP_CONSTEXPR const char CLAP_EXT_NOTE_NAME[] = "clap.note-name";
 
 typedef struct clap_note_name {
-   alignas(1) char   name[CLAP_NAME_SIZE];
-   alignas(4) int32_t port;
-   alignas(4) int32_t key;
-   alignas(4) int32_t channel; // -1 for every channels
+   alignas(1) char name[CLAP_NAME_SIZE];
+   alignas(4) int16_t port;    // -1 for every ports
+   alignas(4) int16_t key;     // -1 for every keys
+   alignas(4) int16_t channel; // -1 for every channels
 } clap_note_name_t;
 
 typedef struct clap_plugin_note_name {
